third_party/WebKit/LayoutTests/fast/css/getComputedStyle/getComputedStyle-exceptions.html

<!doctype html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>
shouldThrow("getComputedStyle()");
shouldThrow("getComputedStyle(document)");
shouldThrow("getComputedStyle(null)");
shouldThrow("getComputedStyle(undefined)");
</script>
</body>
</html>
